ai-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0
numpy==1.26.2